huggingface_downloads/requirements (1).txt
pillow
langchain-google-genai
streamlit==1.29.0
streamlit-chat==0.1.1
streamlit_option_menu==0.3.6
langchain